// src/app/models/TechStack.ts
import type { TechIcon } from "./Icon";

export interface TechStack {
    languages: TechIcon[];
    frameworks: TechIcon[];
    tools: TechIcon[];
    databases: TechIcon[];
    cloud: TechIcon[];
    architectures: TechIcon[];
    messagingProtocols: TechIcon[];
    operatingSystems: TechIcon[];
    devops: TechIcon[];
    testing: TechIcon[];
    platforms: TechIcon[];
}